feat: add new parties (individual & organization)
